requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Supabase Client
supabase==2.0.3

# Pydantic y Settings
pydantic==2.4.2
pydantic-core==2.10.1
pydantic-settings==2.0.3

# Autenticación y Seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.8.0

# Variables de entorno
python-dotenv==1.0.0

# Utilidades
email-validator==2.1.0

# Testing 
pytest==7.4.3
httpx==0.24.1